split custom parser test out
